Avoid the redundant declaration of the function prototypes.
When the server version is 9.4 or later, PG_FUNCTION_INFO_V1 macro
always creates the prototype of its given function, so we avoid
declaring the prototypes of all the global functions for pg_bigm.
IOW, the function prototypes are there only when the server version
is 9.3 or before.

This commit is inspired by the change of pg_trgm:
e7128e8dbb305059c30ec085461297e619bcbff4

Back-patch to 1.1 where 9.4 is supported.

Beena Emerson