Use INT8 to store the byte length of bi-gram string in "bigm" struct.
"bigm" struct has the field "bytelen" which indicates the byte length of
bi-gram string. Previously the data type of that field was INT. This commit
changes the data type to INT8 because the byte length of bi-gram string is
eight bytes at most and INT8 is enough large to indicate that length.
This change is useful to increase the maximum allowed size of the column
indexed by pg_bigm.

This change is cost-effective, i.e., simple but can alleviate the limitation of
the column size very much. So per discussion we decided to back-patch this to
all supported versions.

Beena Emerson, original idea from Sawada Masahiko.