Support triConsistent function for better performance.
The triConsistent function makes the full-text search using
pg_bigm much faster when the search keyword has a rare key.

This commit is inspired by the change of pg_trgm:
97f3014647a5bd570032abd2b809d3233003f13f

This commit bumps the pg_bigm version to 1.2.

Since triConsistent mechanism is available only in PostgreSQL 9.4
or later, this commit drops the support for 9.1, 9.2 and 9.3.
However, per discussion, we concluded to be able to live with
this situation because the current stable version of pg_bigm,
i.e., 1.1, keeps being maintained and supporting those old
PostgreSQL versions.

Reviewed by Sawada Masahiko.