Make gin_bigm_ops use triConsistent function only with PostgreSQL 9.4 or later.
Commit d0b296b introduced triConsistent function which is available
only in 9.4 or later, whatever the server version is. This meant that
pg_bigm 1.2 was not compatible with the server 9.3 or before. We did
this reluctantly because we could not find the way to define
triConsistent function and make gin_bigm_ops use it only when
the server version is 9.4 or later. However we finally found that
we can do that by using DO statement.

This changes the SQL which creates triConsistent function so that
it's executed via DO statement. This DO statement checks the server
version, and then creates the function and registers it to
gin_bigm_ops only when the version is 9.4 or later. Thanks to this
patch, pg_bigm 1.2 is compatible even with 9.3 or before.

Reviewed by Sawada Masahiko