Make pg_gin_pending_stats() check whether given relation is a GIN index.
Previously pg_gin_pending_stats() didn't do such a check, so it could
return an improper and irrelevant result when non-GIN index was specified.
This is a bug. This patch changes pg_gin_pending_stats() so that it exits
with an error when the given relation is not a GIN index.

Also this patch prevents pg_gin_pending_stats() from reading non-local
temporary relations.

Back-patch to all supported versions.