Add pg_gin_pending_cleanup function.
pg_gin_pending_cleanup is the function that cleans up the pending list
of the GIN index by moving tuples in it to the main GIN data structure
in bulk. Even without this function, we can clean up the pending list
by using VACUUM. However, since VACUUM needs to do not only
the pending list cleanup but also other various jobs, it usually takes
a long time and its performance impact is likely to be big.
So pg_gin_pending_cleanup function is useful because we can clean up
the list more quickly and avoid such big performance impact by using
the function.

Reviewed by Masahiko Sawada