Correct documentation for bigm_similarity to indicate case-sensitivity.
The previous documentation for bigm_similarity incorrectly stated that
it was not case-sensitive. This commit updates the documentation to
reflect that bigm_similarity IS case-sensitive.

Please note that the Japanese version of the documentation is not updated
as it already accurately explains bigm_similarity and its case-sensitivity.

Back-patch to pg_bigm 1.1.

Reported and suggested by: Vin Yap
Discussion: https://github.com/pgbigm/pg_bigm/issues/3