Rename html/ directory to docs/ for GitHub Pages compatibility.
To ensure compatibility with GitHub Pages, this commit renames
the html/ directory to docs/. This change is necessary because
GitHub Pages uses the docs/ directory as a default location for
serving web pages. By renaming the directory, the project's
documentation will be accessible on the web when published to
GitHub Pages.