Update README.md to point to GitHub Pages documentation.
This commit updates the README.md file to include a link to
the documentation deployed at GitHub Pages, making it easier
for users to access up-to-date documentation for pg_bigm.
Additionally, it removes a description that suggests an obsolete
official site for pg_bigm, reducing confusion caused by
outdated information. Furthermore, this commit adds copyright
information to the README.md file to ensure proper attribution
for the project.

These changes improve the clarity and accuracy of the project's
documentation and information.